+++ GaluhadPlugins/Pets/Strings/English/Pets.lua
@@ -477,6 +477,19 @@ _PETSTRINGS[393] = {[1]="Proud Lioness";[2]="Available for barter from The City 
 _PETSTRINGS[394] = {[1]="Maned Lion";[2]="Limited time availability: Legendary Item Reward Track Season 10."}
 _PETSTRINGS[395] = {[1]="Wild Chocolate Donkey";[2]="Available for barter during the Midsummer Festival."}
 _PETSTRINGS[396] = {[1]="Wine-red Snake";[2]="Available for barter during the Midsummer Festival."}
+_PETSTRINGS[397] = {[1]="Shadow Hedgehog";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[398] = {[1]="Shadow Cat";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[399] = {[1]="Shadow Raven";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[400] = {[1]="Shadow Lion";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[401] = {[1]="Fierce Hedgehog";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[402] = {[1]="Shadow Nether-hound";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[403] = {[1]="Shadow Spider";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[404] = {[1]="Mirkwood Sabre-cat";[2]="Acquired from the Mirkwood Patron's Luxuriant Coffer."}
+_PETSTRINGS[405] = {[1]="Potato Farmer's Fine Cow";[2]="Available through barter during the Farmers Faire Festival."}
+_PETSTRINGS[406] = {[1]="Golden Drakeling";[2]="Available from participating in community events."}
+_PETSTRINGS[407] = {[1]="Umber Nether-hound";[2]="Available during the Veil of the Nine event on Legendary Worlds."}
+_PETSTRINGS[408] = {[1]="Deep-claw Whelp";[2]="Limited time availability: Legendary Item Reward Track Season 11."}
+_PETSTRINGS[409] = {[1]="Parrot of Darkness";[2]="Limited time availability: Legendary Item Reward Track Season 11."}
 
 _PETSTRINGS[500] = {[1]="Stripey Bat";[2]=""}
 _PETSTRINGS[501] = {[1]="Injured Bat";[2]=""}
@@ -484,37 +497,41 @@ _PETSTRINGS[502] = {[1]="Tan Bat";[2]=""}
 _PETSTRINGS[503] = {[1]="Black-wing Kite";[2]=""}
 _PETSTRINGS[504] = {[1]="Dusk-wing Kite";[2]=""}
 _PETSTRINGS[505] = {[1]="Tan Wolf-dog";[2]=""}
-_PETSTRINGS[506] = {[1]="Umber Nether-hound";[2]=""}
-_PETSTRINGS[507] = {[1]="Cream Corgi";[2]=""}
-_PETSTRINGS[508] = {[1]="Russet Corgi";[2]=""}
-_PETSTRINGS[509] = {[1]="Library Corgi";[2]=""}
-_PETSTRINGS[510] = {[1]="Winter's Wind";[2]=""}
-_PETSTRINGS[511] = {[1]="Rohan Aurochs";[2]=""}
-_PETSTRINGS[512] = {[1]="Deep-claw Whelp";[2]=""}
-_PETSTRINGS[513] = {[1]="Evergreen Sickle-fly";[2]=""}
-_PETSTRINGS[514] = {[1]="Autumn Aurochs";[2]="";}
-_PETSTRINGS[515] = {[1]="Tawny Aurochs";[2]="";}
-_PETSTRINGS[516] = {[1]="Jerboa";[2]="";}
-_PETSTRINGS[517] = {[1]="Striped Jerboa";[2]=""}
-_PETSTRINGS[518] = {[1]="Brown Jerboa";[2]=""}
-_PETSTRINGS[519] = {[1]="Blonde Corgi";[2]=""}
-_PETSTRINGS[520] = {[1]="Jade Speckled Snake";[2]=""}
-_PETSTRINGS[521] = {[1]="Turquoise Snake";[2]=""}
-_PETSTRINGS[522] = {[1]="Pebbled Snake";[2]=""}
-_PETSTRINGS[523] = {[1]="Leafy Green Snake";[2]=""}
-_PETSTRINGS[524] = {[1]="Deep Emerald Snake";[2]="";}
-_PETSTRINGS[525] = {[1]="Fiery Desert Snake";[2]="";}
-_PETSTRINGS[526] = {[1]="Ruby Sand Worm";[2]="";}
-_PETSTRINGS[527] = {[1]="Chestnut Sand Worm";[2]=""}
-_PETSTRINGS[528] = {[1]="Spotted Hamster";[2]=""}
-_PETSTRINGS[529] = {[1]="White Hamster";[2]=""}
-_PETSTRINGS[530] = {[1]="Taupe Hamster";[2]=""}
-_PETSTRINGS[531] = {[1]="Heather Hamster";[2]=""}
-_PETSTRINGS[532] = {[1]="Parrot of the Fiery Sands";[2]=""}
-_PETSTRINGS[533] = {[1]="Water Buffalo Calf";[2]=""}
-_PETSTRINGS[534] = {[1]="Water Buffalo Cow";[2]=""}
-_PETSTRINGS[535] = {[1]="Parrot of Light";[2]=""}
-_PETSTRINGS[536] = {[1]="Water Buffalo Bull";[2]=""}
-_PETSTRINGS[537] = {[1]="Parrot of Darkness";[2]=""}
-_PETSTRINGS[538] = {[1]="Dusky Parrot";[2]=""}
-_PETSTRINGS[539] = {[1]="Dashing Dusky Parrot";[2]=""}
+_PETSTRINGS[506] = {[1]="Cream Corgi";[2]=""}
+_PETSTRINGS[507] = {[1]="Russet Corgi";[2]=""}
+_PETSTRINGS[508] = {[1]="Library Corgi";[2]=""}
+_PETSTRINGS[509] = {[1]="Winter's Wind";[2]=""}
+_PETSTRINGS[510] = {[1]="Rohan Aurochs";[2]=""}
+_PETSTRINGS[511] = {[1]="Evergreen Sickle-fly";[2]=""}
+_PETSTRINGS[512] = {[1]="Autumn Aurochs";[2]=""}
+_PETSTRINGS[513] = {[1]="Tawny Aurochs";[2]=""}
+_PETSTRINGS[514] = {[1]="Jerboa";[2]="";}
+_PETSTRINGS[515] = {[1]="Striped Jerboa";[2]="";}
+_PETSTRINGS[516] = {[1]="Brown Jerboa";[2]="";}
+_PETSTRINGS[517] = {[1]="Blonde Corgi";[2]=""}
+_PETSTRINGS[518] = {[1]="Jade Speckled Snake";[2]=""}
+_PETSTRINGS[519] = {[1]="Turquoise Snake";[2]=""}
+_PETSTRINGS[520] = {[1]="Pebbled Snake";[2]=""}
+_PETSTRINGS[521] = {[1]="Leafy Green Snake";[2]=""}
+_PETSTRINGS[522] = {[1]="Deep Emerald Snake";[2]=""}
+_PETSTRINGS[523] = {[1]="Fiery Desert Snake";[2]=""}
+_PETSTRINGS[524] = {[1]="Ruby Sand Worm";[2]="";}
+_PETSTRINGS[525] = {[1]="Chestnut Sand Worm";[2]="";}
+_PETSTRINGS[526] = {[1]="Spotted Hamster";[2]="";}
+_PETSTRINGS[527] = {[1]="White Hamster";[2]=""}
+_PETSTRINGS[528] = {[1]="Taupe Hamster";[2]=""}
+_PETSTRINGS[529] = {[1]="Heather Hamster";[2]=""}
+_PETSTRINGS[530] = {[1]="Parrot of the Fiery Sands";[2]=""}
+_PETSTRINGS[531] = {[1]="Water Buffalo Calf";[2]=""}
+_PETSTRINGS[532] = {[1]="Water Buffalo Cow";[2]=""}
+_PETSTRINGS[533] = {[1]="Parrot of Light";[2]=""}
+_PETSTRINGS[534] = {[1]="Water Buffalo Bull";[2]=""}
+_PETSTRINGS[535] = {[1]="Dusky Parrot";[2]=""}
+_PETSTRINGS[536] = {[1]="Dashing Dusky Parrot";[2]=""}
+_PETSTRINGS[537] = {[1]="Sunstone Drakeling";[2]=""}
+_PETSTRINGS[538] = {[1]="Amethyst Drakeling";[2]=""}
+_PETSTRINGS[539] = {[1]="Moonstone Drakeling";[2]=""}
+_PETSTRINGS[540] = {[1]="Ruby Drakeling";[2]=""}
+_PETSTRINGS[541] = {[1]="Sapphire Drakeling";[2]=""}
+_PETSTRINGS[542] = {[1]="Ametrine Drakeling";[2]=""}
+_PETSTRINGS[543] = {[1]="Emerald Drakeling";[2]=""}