+++ GaluhadPlugins/Pets/Strings/French/Pets.lua
@@ -477,6 +477,19 @@ _PETSTRINGS[393] = {[1]="Lionne altière";[2]="Peut être échangé auprès de l
 _PETSTRINGS[394] = {[1]="Lion à crinière";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 10."}
 _PETSTRINGS[395] = {[1]="Âne chocolat sauvage";[2]="Disponible par échange au cours du Festival du Solstice d'été."}
 _PETSTRINGS[396] = {[1]="Serpent bordeaux";[2]="Disponible par échange au cours du Festival du Solstice d'été."}
+_PETSTRINGS[397] = {[1]="Hérisson ténébreux";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[398] = {[1]="Chat ténébreux";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[399] = {[1]="Corneille ténébreuse";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[400] = {[1]="Lion ténébreux";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[401] = {[1]="Hérisson redoutable";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[402] = {[1]="Chien ténébreux";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[403] = {[1]="Araignée ténébreuse";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[404] = {[1]="Tigre à dents de sabre de la Forêt Noire";[2]="S'obtient dans le Coffre luxuriant du client de la Forêt Noire."}
+_PETSTRINGS[405] = {[1]="Vache supérieure de récoltant de pommes de terre";[2]="Disponible par échange au cours du Festival de la Foire des fermiers."}
+_PETSTRINGS[406] = {[1]="Petit drake doré";[2]="Disponible en participant à des événements de communauté."}
+_PETSTRINGS[407] = {[1]="Chien couleur terre du tréfonds";[2]="Disponible pendant l'événement Le voile des Neuf sur le serveur légendaire."}
+_PETSTRINGS[408] = {[1]="Jeune griffefosse";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 11."}
+_PETSTRINGS[409] = {[1]="Perroquet des ténèbres";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 11."}
 
 _PETSTRINGS[500] = {[1]="Chauve-souris rayée";[2]=""}
 _PETSTRINGS[501] = {[1]="Chauve-souris blessée";[2]=""}
@@ -484,37 +497,41 @@ _PETSTRINGS[502] = {[1]="Chauve-souris brunie";[2]=""}
 _PETSTRINGS[503] = {[1]="Cerf-volant des ailes noires";[2]=""}
 _PETSTRINGS[504] = {[1]="Cerf-volant des ailes du crépuscule";[2]=""}
 _PETSTRINGS[505] = {[1]="Chien-loup en brun claire";[2]=""}
-_PETSTRINGS[506] = {[1]="Chien couleur terre du tréfonds";[2]=""}
-_PETSTRINGS[507] = {[1]="Corgi crème";[2]=""}
-_PETSTRINGS[508] = {[1]="Corgi brun-roux";[2]=""}
-_PETSTRINGS[509] = {[1]="Corgi de bibliothèque";[2]=""}
-_PETSTRINGS[510] = {[1]="Vent d'hiver";[2]=""}
-_PETSTRINGS[511] = {[1]="Auroch du Rohan";[2]=""}
-_PETSTRINGS[512] = {[1]="Jeune griffefosse";[2]=""}
-_PETSTRINGS[513] = {[1]="Mouche faucheuse tenace";[2]=""}
-_PETSTRINGS[514] = {[1]="Aurochs automnal";[2]="";}
-_PETSTRINGS[515] = {[1]="Aurochs fauve";[2]="";}
-_PETSTRINGS[516] = {[1]="Gerboise";[2]="";}
-_PETSTRINGS[517] = {[1]="Gerboise à rayures";[2]=""}
-_PETSTRINGS[518] = {[1]="Gerboise brune";[2]=""}
-_PETSTRINGS[519] = {[1]="Corgi blond";[2]=""}
-_PETSTRINGS[520] = {[1]="Serpent moucheté de jade";[2]=""}
-_PETSTRINGS[521] = {[1]="Serpent turquoise";[2]=""}
-_PETSTRINGS[522] = {[1]="Serpent tacheté";[2]=""}
-_PETSTRINGS[523] = {[1]="Serpent vert forêt";[2]=""}
-_PETSTRINGS[524] = {[1]="Serpent vert émeraude";[2]="";}
-_PETSTRINGS[525] = {[1]="Serpent ardent du désert";[2]="";}
-_PETSTRINGS[526] = {[1]="Ver des sables vermeil";[2]="";}
-_PETSTRINGS[527] = {[1]="Ver des sables châtain";[2]=""}
-_PETSTRINGS[528] = {[1]="Hamster moucheté";[2]=""}
-_PETSTRINGS[529] = {[1]="Hamster blanc";[2]=""}
-_PETSTRINGS[530] = {[1]="Hamster taupe";[2]=""}
-_PETSTRINGS[531] = {[1]="Hamster mauve";[2]=""}
-_PETSTRINGS[532] = {[1]="Perroquet des sables ardents";[2]=""}
-_PETSTRINGS[533] = {[1]="Bufflon d'eau";[2]=""}
-_PETSTRINGS[534] = {[1]="Bufflonne d'eau";[2]=""}
-_PETSTRINGS[535] = {[1]="Perroquet de lumière";[2]=""}
-_PETSTRINGS[536] = {[1]="Buffle d'eau";[2]=""}
-_PETSTRINGS[537] = {[1]="Perroquet des ténèbres";[2]=""}
-_PETSTRINGS[538] = {[1]="Perroquet mat";[2]=""}
-_PETSTRINGS[539] = {[1]="Perroquet mat fringant";[2]=""}
+_PETSTRINGS[506] = {[1]="Corgi crème";[2]=""}
+_PETSTRINGS[507] = {[1]="Corgi brun-roux";[2]=""}
+_PETSTRINGS[508] = {[1]="Corgi de bibliothèque";[2]=""}
+_PETSTRINGS[509] = {[1]="Vent d'hiver";[2]=""}
+_PETSTRINGS[510] = {[1]="Auroch du Rohan";[2]=""}
+_PETSTRINGS[511] = {[1]="Mouche faucheuse tenace";[2]=""}
+_PETSTRINGS[512] = {[1]="Aurochs automnal";[2]=""}
+_PETSTRINGS[513] = {[1]="Aurochs fauve";[2]=""}
+_PETSTRINGS[514] = {[1]="Gerboise";[2]="";}
+_PETSTRINGS[515] = {[1]="Gerboise à rayures";[2]="";}
+_PETSTRINGS[516] = {[1]="Gerboise brune";[2]="";}
+_PETSTRINGS[517] = {[1]="Corgi blond";[2]=""}
+_PETSTRINGS[518] = {[1]="Serpent moucheté de jade";[2]=""}
+_PETSTRINGS[519] = {[1]="Serpent turquoise";[2]=""}
+_PETSTRINGS[520] = {[1]="Serpent tacheté";[2]=""}
+_PETSTRINGS[521] = {[1]="Serpent vert forêt";[2]=""}
+_PETSTRINGS[522] = {[1]="Serpent vert émeraude";[2]=""}
+_PETSTRINGS[523] = {[1]="Serpent ardent du désert";[2]=""}
+_PETSTRINGS[524] = {[1]="Ver des sables vermeil";[2]="";}
+_PETSTRINGS[525] = {[1]="Ver des sables châtain";[2]="";}
+_PETSTRINGS[526] = {[1]="Hamster moucheté";[2]="";}
+_PETSTRINGS[527] = {[1]="Hamster blanc";[2]=""}
+_PETSTRINGS[528] = {[1]="Hamster taupe";[2]=""}
+_PETSTRINGS[529] = {[1]="Hamster mauve";[2]=""}
+_PETSTRINGS[530] = {[1]="Perroquet des sables ardents";[2]=""}
+_PETSTRINGS[531] = {[1]="Bufflon d'eau";[2]=""}
+_PETSTRINGS[532] = {[1]="Bufflonne d'eau";[2]=""}
+_PETSTRINGS[533] = {[1]="Perroquet de lumière";[2]=""}
+_PETSTRINGS[534] = {[1]="Buffle d'eau";[2]=""}
+_PETSTRINGS[535] = {[1]="Perroquet mat";[2]=""}
+_PETSTRINGS[536] = {[1]="Perroquet mat fringant";[2]=""}
+_PETSTRINGS[537] = {[1]="Petit drake de pierre de soleil";[2]=""}
+_PETSTRINGS[538] = {[1]="Petit drake d'améthyste";[2]=""}
+_PETSTRINGS[539] = {[1]="Petit drake de pierre de lune";[2]=""}
+_PETSTRINGS[540] = {[1]="Petit drake de rubis";[2]=""}
+_PETSTRINGS[541] = {[1]="Petit drake de saphir";[2]=""}
+_PETSTRINGS[542] = {[1]="Petit drake d'amétrine";[2]=""}
+_PETSTRINGS[543] = {[1]="Petit drake d'émeraude";[2]=""}