+++ Changelog.md
@@ -1,4 +1,8 @@
 Change log:
+* 2024-08-01 - U41 - 10 new, 3 newly available, 7 future
+  * Shadow Hedgehog, Shadow Cat, Shadow Raven, Shadow Lion, Fierce Hedgehog, Shadow Nether-hound, Shadow Spider, Mirkwood Sabre-cat, Potato Farmer's Fine Cow, Golden Drakeling
+  * Umber Nether-hound, Deep-claw Whelp, Parrot of Darkness
+  * Sunstone Drakeling, Amethyst Drakeling, Moonstone Drakeling, Ruby Drakeling, Sapphire Drakeling, Ametrine Drakeling, Emerald Drakeling
 * 2024-06-05 - U40.1 - 1 new, 1 newly available
   * Wild Chocolate Donkey
   * Wine-red Snake